requirements.txt
numpy==1.18.2
pandas==1.0.3
tqdm==4.45.0
scikit-learn==0.22.2.post1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
sentencepiece==0.1.83
lightgbm==2.2.3
optuna==1.3.0
tensorflow==1.15.0